queriendo arreglar lo del shift arregle las flechitas del snake

+++ Kernel/include/lib.h
@@ -11,7 +11,7 @@ char *cpuVendor(char *result);
 int getHours();
 int getMins();
 int getSec();
-int getKey();
+uint16_t getKey();
 void outb(uint8_t dest, uint8_t org);
 uint8_t inb(uint8_t dest);
